### 2022-07-20 #4601 @editorialbot
[PRE REVIEW]: POSSA: Power simulation for sequential analyses and multiple hypotheses
**Submitting author:** @gasparl (<a href="http://orcid.org/0000-0001-9401-4830">Gáspár Lukács</a>)
**Repository:** https://github.com/gasparl/possa/
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.5
