### 2022-07-27 #4601 @gasparl
Thank  you @fboehm for taking the editorial role! Below are the reviewers I suggested above, now with names and email addresses – all found on the [joss_reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) list, given in no particular order (actually the reverse order of occurrance on the _joss_reviewers_ list, as suggested above):

- chrisaberson (Chris Aberson; chris.aberson@humboldt.edu)
- sestelo (Marta Sestelo Pérez; sestelo@uvigo.es)
- mingzehuang (Mingze Huang; mingzehuang@gmail.com)
- sdparsons (Sam Parsons; sam.parsons@radboudumc.nl)
- mmrabe (Maximilian Rabe; maximilian.rabe@uni-potsdam.de)
- mcavs (Mustafa Cavus; mustafacavus@eskisehir.edu.tr)
- craddm (Matt Craddock; mcraddock@lincoln.ac.uk)
- cimentadaj (Jorge Cimentada; cimentadaj@gmail.com)
