### 2022-08-02 #4601 @fboehm
@sdparsons @mmrabe @mcavs - would you be willing to review this submission for JOSS? Our reviews are checklist-driven. Please feel free to ask me any questions that you might have. Thanks for considering this!
