### 2022-08-19 gasparl/possa#8 @mingzehuang
Code of Conduct
Hi, @gasparl, you package looks nice! I see there are many package contains a [code of conduct](https://www.contributor-covenant.org). I suggest you may consider to have that:)
