### 2022-08-19 #4643 @mingzehuang
Hi,  @gasparl, your package looks nice:) I see you have "contribution" section in your README file. I would suggest you also include [code of conduct](https://github.com/gasparl/possa/issues/8#issue-1344151572) like many other open source packages:)
