+++ README.md
@@ -41,7 +41,7 @@ For detailed information about each function and parameter, see [the manual](htt
 
 ### Contributions
 
-Given that this project is not part of my "day job", the [intended additional features](https://github.com/gasparl/possa/issues "Issues") grew beyond my capacity. Help would be most welcome; if you are up to it, feel free to [email me](mailto:lkcsgaspar@gmail.com) (or to reply to the relevant issue or open a new one).
+Given that this project is not part of my "day job", the [intended additional features](https://github.com/gasparl/possa/issues "Issues") grew beyond my capacity. Help would be most welcome; if you are up to it, feel free to [email me](mailto:lkcsgaspar@gmail.com) (or to reply to the relevant issue or open a new one; see details [here](https://github.com/gasparl/possa/blob/master/CONTRIBUTING.md "CONTRIBUTING")).
 
 ### Citation
 
