+++ CONTRIBUTING.md
@@ -1,0 +1,35 @@
+# Contributing
+
+Contributions are always welcome; see the open [issues](https://github.com/gasparl/possa/issues "Issues"). When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owner of this repository before making a change.
+
+## All Code Changes Happen Through Pull Requests
+
+1. Fork the repo and create your branch from `master`.
+2. If you've added code that should be tested, add tests.
+3. Make sure the syle of your code is consistent with that of the current one (indentation, etc.).
+4. If you've changed any relevant functionalities, update the documentation.
+5. Ensure the application is working correctly. (Run `devtools::test()` for quick tests, and eventually also `devtools::check(args = '--run-donttest')`.)
+6. Issue that pull request.
+
+## Code of Conduct
+
+Use common sense.
+
+Examples:
+
+* Be respectful of differing viewpoints and experiences
+* Gracefully accept constructive criticism
+* Focus on what is best for the community
+* Have empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* Publishing others' private information without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
+
+## License
+
+By contributing to this repository, you agree that your contributions will be licensed under its BSD-2-Clause License.
