### 2022-08-19 gasparl/possa#8 @gasparl
Now included [here](https://github.com/gasparl/possa/blob/master/CONTRIBUTING.md "CONTRIBUTING").
