### 2022-08-19 #4643 @gasparl
Sure.

- I added [Code of Conduct](https://github.com/gasparl/possa/commit/0d3e4f8a03f6d77ed32c50eb5756fc2de1e3eb6a) together with more details on how contribution works.
- I added a short example to the paper.
