### 2022-08-24 #4643 @mingzehuang
Hi, @gasparl, there is a requirement about automatic test in checklist. Would you mind using [codecov](https://docs.codecov.com/docs) and embedding the badge of coverage rate into your README.md?
