### 2022-08-25 gasparl/possa#9 @mmrabe
JOSS: Conflicting (?) LICENSE files
Your repo contains two LICENSE files, one plain-text (as required by JOSS) and one in Markdown (.md). Do you really need both? If so, could you make sure that the plain-text file is identical to the other?

This relates to JOSS submission openjournals/joss-reviews#4643
