### 2022-08-25 gasparl/possa#9 @gasparl
Hm I remember there was originally (years ago, for another package) some reason for this related to [CRAN](https://cran.r-project.org/) requirements, but I can't recall nor find it.. (Maybe it's not applicable anymore..)
Anyway, I just [corrected](https://github.com/gasparl/possa/commit/f049bd98bf76002dd8e145bbf09164410fc51dd1) this to have one file.
