### 2022-08-25 gasparl/possa#10 @mmrabe
JOSS paper: Add reference
Please add a reference for Pocock's correction in the JOSS paper.

This relates to the JOSS submission openjournals/joss-reviews#4643.
