### 2022-08-25 #4643 @gasparl
> Hi, @gasparl, there is a requirement about automatic test in checklist. Would you mind using [codecov](https://docs.codecov.com/docs) and embedding the badge of coverage rate into your README.md?

Sure, I added it. (The relatively low percentage [52%] is because I included lots of warning messages for making the functions more foolproof, but these don't really need to be tested.)
