### 2022-08-25 gasparl/possa#11 @mmrabe
Suggestion: Extract p-value from common test objects?
An obvious strength of the package is its flexibility to perform power analyses for almost any number of pretty much any statistical test in R. However, this requires that the user code their own test function such as:

```R
customTest = function(sample1) {
  c(
    a = t.test(...)$p.value,
    b = aov(...)[[1]][[1]][["Pr(>F)"]][2]
  )
}
```

As you also point out in your vignettes, this can be somewhat tricky for ANOVA or regression fits but even adding the unusual `$p.value` after the t-test call might be somewhat difficult to understand for the average R user. Would there perhaps be a way of supplying some wrapper function that automatically extracts the desired p-value from common (not all) test objects so that the user can just write something like the following?

```R
customTest = function(sample1) {
  c(
    a = extract(t.test(...)),
    b = extract(aov(...), "FactorA:FactorB")
  )
}
```

This relates to the JOSS submission openjournals/joss-reviews#4643.
