+++ paper/paper.bib
@@ -9,6 +9,21 @@
     doi = {10.1002/ejsp.2023}
 }
 
+@article{pocock:1977,
+    title = {Group sequential methods in the design and analysis of clinical trials},
+    volume = {64},
+    issn = {0006-3444, 1464-3510},
+    url = {https://academic.oup.com/biomet/article-lookup/doi/10.1093/biomet/64.2.191},
+    doi = {10.1093/biomet/64.2.191},
+    language = {en},
+    number = {2},
+    urldate = {2022-08-25},
+    journal = {Biometrika},
+    author = {{Pocock}, S. J.},
+    year = {1977},
+    pages = {191--199},
+}
+
 @article{khandis:2020,
     author = {{Khandis}, R. B. and {Gangestad}, S.},
     title ={On attenuated interactions, measurement error, and statistical power: {Guidelines} for social and personality psychologists},
@@ -21,28 +36,28 @@
 }
 
 @book{cohen:1988,
-	address = {Hillsdale, NJ, US},
-	title = {Statistical power analysis for the behavioral sciences},
-	edition = {2nd ed.},
-	publisher = {Erlbaum},
-	author = {Cohen, J.},
-	year = {1988}
+    address = {Hillsdale, NJ, US},
+    title = {Statistical power analysis for the behavioral sciences},
+    edition = {2nd ed.},
+    publisher = {Erlbaum},
+    author = {Cohen, J.},
+    year = {1988}
 }
 
 @article{perugini:2018,
-	title = {A practical primer to power analysis for simple experimental designs},
-	volume = {31},
-	issn = {2397-8570},
-	url = {http://www.rips-irsp.com/articles/10.5334/irsp.181/},
-	doi = {10.5334/irsp.181},
-	language = {en},
-	number = {1},
-	urldate = {2020-06-23},
-	journal = {International Review of Social Psychology},
-	author = {Perugini, M. and Gallucci, M. and Costantini, G.},
-	month = jul,
-	year = {2018},
-	pages = {20}
+    title = {A practical primer to power analysis for simple experimental designs},
+    volume = {31},
+    issn = {2397-8570},
+    url = {http://www.rips-irsp.com/articles/10.5334/irsp.181/},
+    doi = {10.5334/irsp.181},
+    language = {en},
+    number = {1},
+    urldate = {2020-06-23},
+    journal = {International Review of Social Psychology},
+    author = {Perugini, M. and Gallucci, M. and Costantini, G.},
+    month = jul,
+    year = {2018},
+    pages = {20}
 }
 
 @article{weigl:2020, 
