### 2022-08-25 #4643 @mmrabe
The author @gasparl has written a very useful and flexible tool for simulation-based power analysis. The documentation and vignettes exceed the expected extent by far and provide many helpful examples. The package also already contains tests in the tests/ subdirectory, which I personally feel is probably sufficient. However, I agree with @mingzehuang that it would be good practice to use a service like codecov in addition. The submitted paper has become a lot stronger after adding a simple example, as @mingzehuang suggested.

Before I can check off all items on my checklist and recommend this software paper for publication, please attend to the following:

1. I can not yet check off the license checkbox because the repo contains two license files. Please make sure they match. Issue: gasparl/possa#9
2. In the example in the paper, I recommend that you add a citation for Pocock's correction. Issue: gasparl/possa#10

I would also like to make a minor suggestion, which may make the package more straightforward to use for the average R user and possibly increase the audience:

1. Provide a wrapper function for common test objects that assists the user to extract the p-value. Issue: gasparl/possa#11
