### 2022-08-26 8cc7655
get_p function added
