+++ DESCRIPTION
@@ -1,12 +1,12 @@
 Package: POSSA
 Title: Power Simulation for Sequential Analysis and Multiple Hypotheses
-Version: 0.5.5
-Date: 2022-07-14
+Version: 0.6.0
+Date: 2022-08-26
 Authors@R: c(person("Gáspár", "Lukács", email = "lkcsgaspar@gmail.com", role =
   c("aut", "cre")))
-Description: Calculates, via simulation, power and appropriate "stopping"
+Description: Calculates, via simulation, power and appropriate stopping
   alpha boundaries (and/or futility bounds) for sequential analyses (i.e.,
-  "group sequential design") as well as for multiple hypotheses (multiple tests
+  group sequential design) as well as for multiple hypotheses (multiple tests
   included in an analysis), given any specified global error rate. This enables
   the sequential use of practically any significance test, as long as the
   underlying data can be simulated in advance to a reasonable approximation.
