+++ NAMESPACE
@@ -1,8 +1,12 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(get_p,aov)
+S3method(get_p,aovlist)
+S3method(get_p,default)
 S3method(print,possa_pow_df)
 S3method(print,possa_pow_list)
 S3method(print,possa_sim_df)
+export(get_p)
 export(pow)
 export(sim)
 import(data.table)
