### 2022-08-26 gasparl/possa#11 @gasparl
Okay, I now [included](https://github.com/gasparl/possa/commit/8cc765587efa8633c2a0ccb5f6ab6a02f85e7f9d#diff-f0344f97a72f579bfdfdbbef0ba51b1659a5ece74fa262e130fcf9b0455ab095) a function for this. Currently, it only supports R `stats` functions (including the particularly complicated `aov`). It'd of course be hassle to do it for each test separately, but I guess anyone who is more familiar with specific tests could fairly easily contribute with more, now that the template is there.

I might note however two drawbacks (for future considerations):
a) I realize this is not usual or likely (especially for R `stats`), but even so, if anything changes in other packages, it might take the wrong value (e.g., a different effect in the ANOVA) without the user noticing, which would be extremely problematic. Of course, this could be tested via the automated tests, but that doesn't at all guarantee the corrections will be done in time, let alone installed by users. Anyway, I included cautionary notes in the function description.
b) At tens of thousands of iterations, this could add a little time (though not much, in my `aov` testings it added like 30 sec at 45000 iterations).
