### 2022-08-26 #4643 @gasparl
I now [added](https://github.com/gasparl/possa/commit/8cc765587efa8633c2a0ccb5f6ab6a02f85e7f9d) the p value extraction function as [suggested](https://github.com/gasparl/possa/issues/11) by @mmrabe.

@fboehm, this concludes my replies/revisions to all reviewer requests and comments.
