### 2022-08-26 gasparl/possa#11 @mmrabe
Wonderful! I agree, this can be dangerous. It baffles me that there is no standard way of extracting p-values from statistical models in R, in contrast to coefficients for example. I do appreciate the your effort of including such a method, which may make the construction of the test function easier for some users.
