### 2022-08-26 #4643 @mmrabe
I confirm that the new ``get_p(...)`` method is working and well documented. It has even found its way into the vignettes already. Thanks for adding this right away, @gasparl! I believe this will be very helpful to some potential users. I can without any concern recommend this paper for publication in JOSS. Great job!
