### 2022-08-26 #4643 @fboehm
The paper looks really good. I have a minor correction that is needed:

line 50 in the pdf: capitalize R
One way to do this is to put braces around capital R in your bib file: {R}

I have no other edits to suggest for the paper
