### 2022-08-26 #4643 @fboehm
I manually checked the currently present DOIs, and they all resolved to the intended targets.
