+++ paper/paper.bib
@@ -41,7 +41,8 @@
     edition = {2nd ed.},
     publisher = {Erlbaum},
     author = {Cohen, J.},
-    year = {1988}
+    year = {1988},
+    doi = {10.4324/9780203771587}
 }
 
 @article{perugini:2018,
@@ -61,7 +62,7 @@
 }
 
 @article{weigl:2020, 
-    title={Group Sequential Designs Applied in Psychological Research}, 
+    title={Group sequential designs applied in psychological research}, 
     volume={16}, 
     doi={10.5964/meth.2811}, 
     number={1}, 
@@ -74,7 +75,7 @@
 
 
 @article{green:2016,
-    title = {simr: an R package for power analysis of generalised linear mixed models by simulation},
+    title = {{SIMR}: An {R} package for power analysis of generalised linear mixed models by simulation},
     author = {Peter Green and Catriona J. MacLeod},
     journal = {Methods in Ecology and Evolution},
     volume = {7},
@@ -93,12 +94,12 @@
     number = {1},
     pages = {251524592095150},
     author = {Lakens, D. and Caldwell, A.},
-    title = {Simulation-Based Power Analysis for Factorial Analysis of Variance Designs},
+    title = {Simulation-based power analysis for factorial analysis of variance designs},
     journal = {Advances in Methods and Practices in Psychological Science},
 }
 
 @manual{hughes:2017,
-    title = {paramtest: Run a Function Iteratively While Varying Parameters},
+    title = {{paramtest}: Run a function iteratively while varying parameters},
     author = {Jeffrey Hughes},
     year = {2017},
     note = {R package version 0.1.0},
@@ -106,7 +107,7 @@
 }
 
 @manual{wassmer:2022,
-    title = {rpact: Confirmatory Adaptive Clinical Trial Design and Analysis},
+    title = {{rpact}: Confirmatory adaptive clinical trial design and analysis},
     author = {Gernot Wassmer and Friedrich Pahlke},
     year = {2022},
     note = {R package version 3.2.3},
@@ -114,7 +115,7 @@
 }
 
 @manual{anderson:2022,
-    title = {gsDesign: Group Sequential Design},
+    title = {gsDesign: Group sequential design},
     author = {Keaven Anderson},
     year = {2022},
     note = {R package version 3.2.2},
@@ -122,8 +123,8 @@
 }
 
 @manual{pahl:2022,
-    title = {GroupSeq: Group Sequential Design Probabilities - With Graphical User
-    Interface},
+    title = {GroupSeq: Group sequential design probabilities - with graphical user
+    interface},
     author = {Roman Pahl},
     year = {2022},
     note = {R package version 1.4.0},
