### 2022-08-27 #4643 @gasparl
- Thank you for noticing @fboehm, I [corrected](https://github.com/gasparl/possa/commit/cbd18e4e4f202eff661d9ea09525807f73bbfc4d#diff-c7ec3d0f2ff5bf83e6afaac85d8fa2a792d68b694676683802edf49098a891d3R75) to `{R}`, as well as a few other lettercasing issues among the references (R package names to correspond to their original paper/CRAN titles).
- The auto-suggested doi is not the right one (wrong edition), but I found the [right one](https://doi.org/10.4324/9780203771587) and added it.
