### 2022-08-27 #4643 @fboehm
The references all look good now. For the next steps, @gasparl, we need you to make a new release of the package and archive it, for example, with zenodo.org. Once you complete those tasks, please report here the version number and doi. Please ensure that the archive's author names and title match exactly those of the paper.pdf.
