### 2022-08-27 #4643 @gasparl
Sure @fboehm, here it is:
Version number: v0.6.1
DOI: https://doi.org/10.5281/zenodo.7027767

(The Zenodo badge with permanent DOI is also displayed on the GitHub [repo](https://github.com/gasparl/possa#citation).)
